Move unit tests from shipit_uplift to shipit_code_coverage_backend
