--- src/shipit_code_coverage_backend/tests/test_example.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-def test_example(app):
-    assert True
